Moved types to FrequentQuestions.vue. Surprisingly only Answer needs to be exported. The other two were implicit exports?
